fix: functional pagination hook
